Refactor vmsctl interface.
Now it is a class,a nd you call methods on it to have vmsctl actions happen.
Will follow up with a set of convenient shorthands.

Note that the class constructor takes as input the instance ID, rather than an
object encapsulating the whole instance. This is on purpose because such
objects are not known to the harness (they could be...)