Renamed pytest.py to py.test
Having a file called pytest.py in openstack-test precluded the use of a natively
installed py.test because the local pytest.py would always be imported. It's
handy to be able to use your own install of py.test because of test-running
plugins, like xdist, that you might have installed.