Updating instance iptables retrieval
Subsitute any of the host ips with the string HOST_IP to ensure
that comparing iptables rules works when the instances are on
different hosts.

Also updated how the server id is extracted from the server obj.
In Essex the id is never returned by the API so we need to
determine it using the instance name (i.e. instance-XXX, where
XXX is the id in hex format).