Large refactor of openstack-test for general consumption.
There were many assumptions baked into the test framework about
servers, keys, paths, hosts, etc. I've now refactored this so that
this can be run on any cluster but it may require quite a few more
command line parameters. This is part of an effort to fire-off one
time use test clusters for integration purposes, or to ship this
test repository to other public users.