Implement tests for bless with volume.
Teach grinder about the cobalt volume support capability.

Massive sed of the client to represent that it's a nova client, different from
a cinder client.

Refactor instance creation to fix the multi-instance launch case.  We were only
waiting for boot on the first instance. Return a list of grinder Instance
objects for this case, which will come in handy for volume tests.

Create separate module for volume tests, as well as separate volume object for
'with' contexts.

Change-Id: I023b1506342a41e220021d6dbe6bdd2fe43442a9