Added wait condition for snapshots and volumes (GRIND-10)
When a live image is discarded, all volume snapshots associated with
it will also get deleted. Volumes can only be deleted once these
snapshots are completely gone, so this extra wait is required.

When a cloned instance is deleted, all volumes associated with it
will automatically be deleted. An extra wait is performed for
these volumes.

Change-Id: I444f58e7ac39da07012952424536b88dcdc2a4f9