Refactoring how grinder used lockfiles to serialize policy tests.
The previous implementation of lockfiles was racy because every tests
was creating and deleting the lockfile, which lead to the possibility
of multiple tests using a different file on the same path. Now we
create the lockfile during test configuration and all grinder
processes reuses the fd to ensure they all point to the same lock
file.

Addresses GRIND-29.

Change-Id: I1d29f2f727c47786e29b05f429d843082d41b261